Checksums of raw data coming from sensors
Checksums of:
NMEA data
pressure data
imu data
button data

They are recorded to indicate progress of collecting of data over the time.
Directories with names that start with 00_ are considered to contain preparatory data, not to be used in "production". The data has been used e.g. to create more correct route map.
Directories with names that start with prefixes other than 00_ contain data that will be used for "production". The shape of routes is more or less established, the number and location of waypoints is more or less established.